Changes to rooms components and name fixes
Change AddRooms.js to CreateRoom.js
Added alert on submit
Added number validation to capacity field
Move table from CreateRoom.js to RoomList.js
Change rooms.js to room.js
